third_party/WebKit/LayoutTests/fast/dom/custom/exception-from-constructor-expected.txt

Tests throwing an exception during registration of constructor

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.registerElement("x-a", {prototype: proto}) threw exception Exception thrown from getter.
PASS successfullyParsed is true

TEST COMPLETE
